Consolidate duplicate DataSource model files
