+++ server/src/models/datasource.model.ts
@@ -1,8 +1,86 @@
 import mongoose, { Schema, Document } from 'mongoose';
-import { DataSourceDocument } from '../types/inventory';
 
-// Create the DataSource schema
-const DataSourceSchema = new Schema<DataSourceDocument>({
+// Define the DataSource document interface
+export interface DataSourceDocument extends Document {
+  name: string;
+  type: 'county-website' | 'state-records' | 'tax-database' | 'api' | 'pdf';
+  url: string;
+  region: {
+    state: string;
+    county?: string;
+  };
+  collectorType: string;
+  schedule: {
+    frequency: 'hourly' | 'daily' | 'weekly' | 'monthly' | 'manual';
+    dayOfWeek?: number;
+    dayOfMonth?: number;
+  };
+  metadata?: {
+    lookupMethod?: 'account_number' | 'parcel_id';
+    selectors?: any;
+    lienUrl?: string;
+    [key: string]: any;
+  };
+  status: 'active' | 'inactive' | 'error';
+  lastCollected?: Date;
+  errorMessage?: string;
+  createdAt: Date;
+  updatedAt: Date;
+}
+
+// For backward compatibility with code using IDataSource
+export interface IDataSource extends DataSourceDocument {}
+
+// Define the region schema
+const RegionSchema = new Schema({
+  state: {
+    type: String,
+    required: true,
+    trim: true,
+  },
+  county: {
+    type: String,
+    trim: true,
+  },
+}, { _id: false });
+
+// Define the schedule schema
+const ScheduleSchema = new Schema({
+  frequency: {
+    type: String,
+    required: true,
+    enum: ['hourly', 'daily', 'weekly', 'monthly', 'manual'],
+    default: 'manual',
+  },
+  dayOfWeek: {
+    type: Number,
+    min: 0,
+    max: 6,
+  },
+  dayOfMonth: {
+    type: Number,
+    min: 1,
+    max: 31,
+  },
+}, { _id: false });
+
+// Define the metadata schema
+const MetadataSchema = new Schema({
+  lookupMethod: {
+    type: String,
+    enum: ['account_number', 'parcel_id'],
+  },
+  selectors: {
+    type: Schema.Types.Mixed,
+  },
+  lienUrl: {
+    type: String,
+    trim: true,
+  },
+}, { _id: false, strict: false });
+
+// Define the DataSource schema
+const DataSourceSchema = new Schema({
   name: {
     type: String,
     required: true,
@@ -19,15 +97,8 @@ const DataSourceSchema = new Schema<DataSourceDocument>({
     trim: true,
   },
   region: {
-    state: {
-      type: String,
-      required: true,
-      trim: true,
-    },
-    county: {
-      type: String,
-      trim: true,
-    },
+    type: RegionSchema,
+    required: true,
   },
   collectorType: {
     type: String,
@@ -35,35 +106,14 @@ const DataSourceSchema = new Schema<DataSourceDocument>({
     trim: true,
   },
   schedule: {
-    frequency: {
-      type: String,
-      required: true,
-      enum: ['hourly', 'daily', 'weekly', 'monthly', 'manual'],
-      default: 'daily',
-    },
-    dayOfWeek: {
-      type: Number,
-      min: 0,
-      max: 6,
-    },
-    dayOfMonth: {
-      type: Number,
-      min: 1,
-      max: 31,
-    },
+    type: ScheduleSchema,
+    default: () => ({
+      frequency: 'manual'
+    }),
   },
   metadata: {
-    lookupMethod: {
-      type: String,
-      enum: ['account_number', 'parcel_id'],
-    },
-    selectors: {
-      type: Schema.Types.Mixed,
-    },
-    lienUrl: {
-      type: String,
-      trim: true,
-    },
+    type: MetadataSchema,
+    default: () => ({}),
   },
   status: {
     type: String,
@@ -78,7 +128,7 @@ const DataSourceSchema = new Schema<DataSourceDocument>({
     type: String,
   },
 }, {
-  timestamps: true
+  timestamps: true,
 });
 
 // Create indexes for common search fields
@@ -89,4 +139,7 @@ DataSourceSchema.index({ 'region.state': 1, 'region.county': 1 });
 DataSourceSchema.index({ status: 1 });
 
 // Create and export the model
-export const DataSource = mongoose.model<DataSourceDocument>('DataSource', DataSourceSchema); 
+export const DataSource = mongoose.model<DataSourceDocument>('DataSource', DataSourceSchema);
+
+// For backward compatibility with code using default export
+export default DataSource;
