+++ server/src/models/index.ts
@@ -3,4 +3,5 @@ export { Property } from './property.model';
 export { State } from './state.model';
 export { County } from './county.model';
 export { USMap } from './usmap.model';
-export { Controller } from './controller.model'; 
+export { Controller } from './controller.model';
+export { DataSource, IDataSource, DataSourceDocument } from './datasource.model';
