+++ server/src/routes/data-source.ts
@@ -1,5 +1,5 @@
 import express from 'express';
-import DataSource from '../models/DataSource';
+import { DataSource } from '../models/datasource.model';
 import { collectorManager } from '../services/dataCollection/CollectorManager';
 import logger from '../utils/logger';
